result: extract constructor logic to query_result.rs
The logic will become much more complex once we bump to 0.15.
It introduces additional nesting, thus I prefer to extract it to separate
functions (since cass_session_execute is already complex enough).

I also moved the necessary utility functions to query_result module.
